requirements.txt
python-telegram-bot
fastapi
uvicorn
sqlalchemy
psycopg2
pandas
langchain
mcp
langchain_openai
langchain_core
langchain_community
langchain_mcp_adapters 
langgraph
python-dotenv